Algumas rotas agora só podem ser acessadas por adm logados

+++ src/controllers/userController/UserControllerDelete.js
@@ -3,6 +3,10 @@ require("../../db/connection");
 //deletar usuarios que nao sao adms
 const deleteUser = async (req, res) => {
     try {
+        const userPermission = req.userPermission;
+        if(userPermission == 0){
+            return res.status(401).json({ message: "Não autorizado a acessar esta rota."})
+        }
         const deleteUser = await User.deleteMany({ profile: 'user'});
         if (deleteUser.deletedCount === 0) {
             return res.json({ message: 'Não há usuários para deletar' });
@@ -16,6 +20,10 @@ const deleteUser = async (req, res) => {
 //deletar todos os usuarios
 const deleteAllData = async (req, res) => {
     try {
+        const userPermission = req.userPermission;
+        if(userPermission == 0){
+            return res.status(401).json({ message: "Não autorizado a acessar esta rota."})
+        }
         const deleteUsers = await User.deleteMany({})
         if (deleteUsers.deletedCount === 0) {
             return res.json({ message: 'Não há usuários para deletar' });
